7th commit.

+++ myFirstGame/src/com/tutorial/game/FastEnemy.java
@@ -24,19 +24,19 @@ public class FastEnemy extends GameObject {
 		if(y<= 0 || y>= Game.HEIGHT-32) velY*= -1;
 		if(x<= 0 || x>= Game.WIDTH-16) velX*= -1;
 		
-		 handler.addObject(new Trail(x,y,ID.Trail, Color.cyan, 16, 16, 0.02f, handler));
+		 handler.addObject(new Trail((int)x,(int)y,ID.Trail, Color.cyan, 16, 16, 0.02f, handler));
 	}
 
 	@Override
 	public void render(Graphics g) {
 		g.setColor(Color.cyan);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x,(int) y, 16, 16);
 		
 		
 	}
 	 public Rectangle getBounds() {
 		
-			return new Rectangle(x,y, 16,16);
+			return new Rectangle((int)x,(int)y, 16,16);
 	}
 
 }
