+++ myFirstGame/src/com/tutorial/game/HUD.java
@@ -6,9 +6,9 @@ import java.awt.Graphics;
 //Heads up display
 public class HUD {
 	
-	public static int HEALTH = 100;
+	public static float HEALTH = 100;
 	
-	private int greenValue =255;
+	private float greenValue =255;
 	
 	private int score = 0;
 	private int level = 1;
@@ -26,11 +26,11 @@ public class HUD {
 		g.setColor(Color.gray);
 		g.fillRect(15, 15, 200, 32);
 		
-		g.setColor(new Color(75,greenValue, 0));
-		g.fillRect(15, 15, HEALTH*2, 32);
+		g.setColor(new Color(75,(int)greenValue, 0));
+		g.fillRect(15,15, (int)HEALTH*2, 32);
 		
 		g.setColor(Color.white);
-		g.drawRect(15, 15, 200, 32);
+		g.drawRect(15, (int) 15, (int) 200,(int)  32);
 		
 		g.drawString("SCORE :"+ score, 15, 64);
 		g.drawString("LEVEL :"+ level, 15, 80);
